+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Changed
 
+  - Session - Sort groups and remove prefix in authentication menu.
   - Improve contrast for selected item in sidebar menu.
   - Show not found message for NuvlaBox
-
 
 ## [2.19.0] - 2021-06-04